refactor: implement evaluation architecture with template method and strategy patterns Architecture: - Template Method pattern: BaseEvaluator defines common workflow, subclasses override hooks - Strategy pattern: EvaluatorFactory centralizes evaluator instantiation by stage - Reorganized factories with unified experiment generation interfaces Changes: - app/core/stages/evaluation/: Added base_evaluator, evaluator_factory, specialized evaluators - app/core/stages/factories/: Unified factory pattern, moved from legacy registry - ModelRegistry: JSON serialization, feature preservation, atomic versioning Model Selection Stage: - Top-3 model extraction ensuring family diversity (RandomForest, SVM, LogisticRegression) - Metadata storage: top_k_models_by_family, all_models_by_family for auditing - Consistent metric sorting and result ranking Testing: - 70 comprehensive tests for evaluation subsystem - Template method, factory, and stage-specific evaluator coverage - Edge cases: single family, insufficient families, many families - Real pipeline integration with sklearn All tests passing.

+++ app/core/domain/experiments/experiment.py
@@ -72,7 +72,7 @@ class Experiment:
 
             if metric_name.startswith("train_") or metric_name.startswith("test_"):
 
-                mean_value = float(np.mean(values))
+                mean_value = (np.mean(values))
 
                 # Fix sklearn negative regression metrics
                 if metric_name.endswith(("neg_mean_squared_error",
